fixup??? compat/mingw: split out POSIX-related bits
A simple `#include "posix.h"` without dropping `-Icompat` from the
command-line will find `compat/posix.h`. In
`compat/mingw/compat-util.h`, that is not what we want to include, but
`compat/mingw/posix.h`.

In the long run we may want to reconsider using identical basenames
multiple times. This also helps with `gdb` where it is hard to
impossible to set breakpoints or list code in files with files with
filenames that are reused in other directories.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>